llama: factor out model loader
