+++ src/llama-mmap.cpp
@@ -1,0 +1,557 @@
+#include "llama-mmap.h"
+#include "llama-impl.h"
+#include "ggml.h"
+
+#include <stdexcept>
+#include <cstring>
+
+#if defined(_WIN32)
+
+static std::string llama_format_win_err(DWORD err) {
+    LPSTR buf;
+    size_t size = FormatMessageA(FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS,
+                                 NULL, err, MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT), (LPSTR)&buf, 0, NULL);
+    if (!size) {
+        return "FormatMessageA failed";
+    }
+    std::string ret(buf, size);
+    LocalFree(buf);
+    return ret;
+}
+
+std::string llama_file::GetErrorMessageWin32(DWORD error_code) const {
+    std::string ret;
+    LPSTR lpMsgBuf = NULL;
+    DWORD bufLen = FormatMessageA(FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS,
+            NULL, error_code, MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT), (LPSTR)&lpMsgBuf, 0, NULL);
+    if (!bufLen) {
+        ret = format("Win32 error code: %s", error_code);
+    } else {
+        ret = lpMsgBuf;
+        LocalFree(lpMsgBuf);
+    }
+
+    return ret;
+}
+
+llama_file::llama_file(const char * fname, const char * mode) {
+    fp = ggml_fopen(fname, mode);
+    if (fp == NULL) {
+        throw std::runtime_error(format("failed to open %s: %s", fname, strerror(errno)));
+    }
+    fp_win32 = (HANDLE) _get_osfhandle(_fileno(fp));
+    seek(0, SEEK_END);
+    size = tell();
+    seek(0, SEEK_SET);
+}
+
+size_t llama_file::tell() const {
+    // SetFilePointerEx returns the current position when seeking relative 0 bytes
+    LARGE_INTEGER li;
+    li.QuadPart = 0;
+    BOOL ret = SetFilePointerEx(fp_win32, li, &li, FILE_CURRENT);
+    if (!ret) {
+        throw std::runtime_error(format("read error: %s", GetErrorMessageWin32(GetLastError()).c_str()));
+    }
+
+    return li.QuadPart;
+}
+
+void llama_file::seek(size_t offset, int whence) const {
+    // no need to convert SEEK_* to FILE_*. The enums are the same.
+    // Still, keep static asserts to avoid failures in the future.
+    static_assert(SEEK_SET == FILE_BEGIN, "SEEK_SET != FILE_BEGIN");
+    static_assert(SEEK_CUR == FILE_CURRENT, "SEEK_CUR != FILE_CURRENT");
+    static_assert(SEEK_END == FILE_END, "SEEK_END != FILE_END");
+
+    LARGE_INTEGER li;
+    li.QuadPart = offset;
+    BOOL ret = SetFilePointerEx(fp_win32, li, NULL, whence);
+    if (!ret) {
+        throw std::runtime_error(format("read error: %s", GetErrorMessageWin32(GetLastError()).c_str()));
+    }
+}
+
+void llama_file::read_raw(void * ptr, size_t len) const {
+    // On Win32 ReadFile is significant faster than fread which is again significant faster than std::fstream. Thus
+    // use the Win32 API to do file io instead of the C/C++ library functions.
+
+    // There are conditions under which ReadFile cannot read chunks >64MB.
+    // Thus split the operation into smaller chunks if len exceeds this limit.
+    size_t bytes_read = 0;
+    while (bytes_read < len) {
+        size_t chunk_size = std::min<size_t>(len - bytes_read, 64*1024*1024);
+        DWORD chunk_read = 0;
+        BOOL result = ReadFile(fp_win32, reinterpret_cast<char*>(ptr) + bytes_read, chunk_size, &chunk_read, NULL);
+        if (!result) {
+            throw std::runtime_error(format("read error: %s", GetErrorMessageWin32(GetLastError()).c_str()));
+        }
+        if (chunk_read < chunk_size || chunk_read == 0) {
+            throw std::runtime_error("unexpectedly reached end of file");
+        }
+
+        bytes_read += chunk_read;
+    } ;
+}
+
+void llama_file::write_raw(const void * ptr, size_t len) const {
+    // There are conditions under which WriteFile cannot write chunks >64MB.
+    // Thus split the operation into smaller chunks if len exceeds this limit.
+    size_t bytes_written = 0;
+    while (bytes_written < len) {
+        size_t chunk_size = std::min<size_t>(len - bytes_written, 64*1024*1024);
+        DWORD chunk_written = 0;
+        BOOL result = WriteFile(fp_win32, reinterpret_cast<char const*>(ptr) + bytes_written, chunk_size, &chunk_written, NULL);
+        if (!result) {
+            throw std::runtime_error(format("write error: %s", GetErrorMessageWin32(GetLastError()).c_str()));
+        }
+        if (chunk_written < chunk_size || chunk_written == 0) {
+            throw std::runtime_error("unexpectedly failed to write bytes");
+        }
+
+        bytes_written += chunk_written;
+    }
+}
+
+llama_file::~llama_file() {
+    if (fp) {
+        std::fclose(fp);
+    }
+}
+
+#else
+
+llama_file::llama_file(const char * fname, const char * mode) {
+    fp = ggml_fopen(fname, mode);
+    if (fp == NULL) {
+        throw std::runtime_error(format("failed to open %s: %s", fname, strerror(errno)));
+    }
+    seek(0, SEEK_END);
+    size = tell();
+    seek(0, SEEK_SET);
+}
+
+size_t llama_file::tell() const {
+#ifdef _WIN32
+    __int64 ret = _ftelli64(fp);
+#else
+    long ret = std::ftell(fp);
+#endif
+    if (ret == -1) {
+        throw std::runtime_error(format("ftell error: %s", strerror(errno)));
+    }
+
+    return (size_t) ret;
+}
+
+void llama_file::seek(size_t offset, int whence) const {
+#ifdef _WIN32
+    int ret = _fseeki64(fp, (__int64) offset, whence);
+#else
+    int ret = std::fseek(fp, (long) offset, whence);
+#endif
+    if (ret != 0) {
+        throw std::runtime_error(format("seek error: %s", strerror(errno)));
+    }
+}
+
+void llama_file::read_raw(void * ptr, size_t len) const {
+    if (len == 0) {
+        return;
+    }
+    errno = 0;
+    std::size_t ret = std::fread(ptr, len, 1, fp);
+    if (ferror(fp)) {
+        throw std::runtime_error(format("read error: %s", strerror(errno)));
+    }
+    if (ret != 1) {
+        throw std::runtime_error("unexpectedly reached end of file");
+    }
+}
+
+void llama_file::write_raw(const void * ptr, size_t len) const {
+    if (len == 0) {
+        return;
+    }
+    errno = 0;
+    size_t ret = std::fwrite(ptr, len, 1, fp);
+    if (ret != 1) {
+        throw std::runtime_error(format("write error: %s", strerror(errno)));
+    }
+}
+
+llama_file::~llama_file() {
+    if (fp) {
+        std::fclose(fp);
+    }
+}
+#endif
+using llama_files = std::vector<std::unique_ptr<llama_file>>;
+
+#ifdef _POSIX_MAPPED_FILES
+
+llama_mmap::llama_mmap(struct llama_file * file, size_t prefetch, bool numa, [[maybe_unused]] bool use_thp) {
+    size = file->size;
+    int fd = fileno(file->fp);
+    int flags = MAP_SHARED;
+    // prefetch/readahead impairs performance on NUMA systems
+    if (numa)  { prefetch = 0; }
+#ifdef __linux__
+    // advise the kernel to read the file sequentially (increases readahead)
+    if (posix_fadvise(fd, 0, 0, POSIX_FADV_SEQUENTIAL)) {
+        LLAMA_LOG_WARN("warning: posix_fadvise(.., POSIX_FADV_SEQUENTIAL) failed: %s\n",
+                strerror(errno));
+    }
+    if (prefetch) { flags |= MAP_POPULATE; }
+    if (use_thp) {
+        size_t huge = get_default_huge_page_size();
+        auto size = huge*((file->size + huge - 1)/huge);
+        addr = mmap(nullptr, size, PROT_READ | PROT_WRITE, MAP_PRIVATE | MAP_ANONYMOUS | MAP_HUGETLB, -1, 0);
+        if (addr != MAP_FAILED) {
+            printf("%s: using THP with page size %zu MiB ", __func__, huge/(1024*1024));
+            fflush(stdout);
+            size_t tot = 0;
+            while (tot < file->size) {
+                auto n_read = pread(fd, static_cast<char*>(addr) + tot, file->size - tot, tot);
+                if (n_read < 0) throw std::runtime_error(format("Reading into mapped huge pages failed at %zu (%s)", tot, strerror(errno)));
+                printf(".");  fflush(stdout);
+                tot += n_read;
+            }
+            printf(" done\n");
+            mapped_fragments.emplace_back(0, file->size);
+            mapped_page_size = huge;
+            return;
+        }
+        else {
+            fprintf(stderr, "%s: mmap with huge page size %zu MiB failed (%s)\n", __func__, huge/(1024*1024), strerror(errno));
+        }
+    }
+#endif
+    addr = mmap(NULL, file->size, PROT_READ, flags, fd, 0);
+    if (addr == MAP_FAILED) { // NOLINT
+        throw std::runtime_error(format("mmap failed: %s", strerror(errno)));
+    }
+
+    if (prefetch > 0) {
+        // advise the kernel to preload the mapped memory
+        if (posix_madvise(addr, std::min(file->size, prefetch), POSIX_MADV_WILLNEED)) {
+            LLAMA_LOG_WARN("warning: posix_madvise(.., POSIX_MADV_WILLNEED) failed: %s\n",
+                    strerror(errno));
+        }
+    }
+    if (numa) {
+        // advise the kernel not to use readahead
+        // (because the next page might not belong on the same node)
+        if (posix_madvise(addr, file->size, POSIX_MADV_RANDOM)) {
+            LLAMA_LOG_WARN("warning: posix_madvise(.., POSIX_MADV_RANDOM) failed: %s\n",
+                    strerror(errno));
+        }
+    }
+
+    // initialize list of mapped_fragments
+    mapped_fragments.emplace_back(0, file->size);
+}
+
+static void llama_mmap::align_range(size_t * first, size_t * last, size_t page_size) {
+    // align first to the next page
+    size_t offset_in_page = *first & (page_size - 1);
+    size_t offset_to_page = offset_in_page == 0 ? 0 : page_size - offset_in_page;
+    *first += offset_to_page;
+
+    // align last to the previous page
+    *last = *last & ~(page_size - 1);
+
+    if (*last <= *first) {
+        *last = *first;
+    }
+}
+
+// partially unmap the file in the range [first, last)
+void llama_mmap::unmap_fragment(size_t first, size_t last) {
+    // note: this function must not be called multiple times with overlapping ranges
+    // otherwise, there is a risk of invalidating addresses that have been repurposed for other mappings
+    int page_size = mapped_page_size > 0 ? mapped_page_size : sysconf(_SC_PAGESIZE);
+    align_range(&first, &last, page_size);
+    size_t len = last - first;
+
+    if (len == 0) {
+        return;
+    }
+
+    GGML_ASSERT(first % page_size == 0);
+    GGML_ASSERT(last % page_size == 0);
+    GGML_ASSERT(last > first);
+
+    void * next_page_start = (uint8_t *) addr + first;
+
+    // unmap the range
+    if (munmap(next_page_start, len)) {
+        LLAMA_LOG_WARN("warning: munmap failed: %s\n", strerror(errno));
+    }
+
+    // update the list of mapped fragments to avoid unmapping the same range again in the destructor
+    std::vector<std::pair<size_t, size_t>> new_mapped_fragments;
+    for (const auto & frag : mapped_fragments) {
+        if (frag.first < first && frag.second > last) {
+            // the range is in the middle of the fragment, split it
+            new_mapped_fragments.emplace_back(frag.first, first);
+            new_mapped_fragments.emplace_back(last, frag.second);
+        } else if (frag.first < first && frag.second > first) {
+            // the range starts in the middle of the fragment
+            new_mapped_fragments.emplace_back(frag.first, first);
+        } else if (frag.first < last && frag.second > last) {
+            // the range ends in the middle of the fragment
+            new_mapped_fragments.emplace_back(last, frag.second);
+        } else if (frag.first >= first && frag.second <= last) {
+            // the range covers the entire fragment
+        } else {
+            // the range is outside the fragment
+            new_mapped_fragments.push_back(frag);
+        }
+    }
+    mapped_fragments = std::move(new_mapped_fragments);
+}
+
+#ifdef __linux__
+static int llama_mmap::get_default_huge_page_size() {
+    int pg_size = 2048;
+    std::ifstream in("/proc/meminfo");
+    if (in) {
+        std::string line;
+        while (true) {
+            std::getline(in, line);
+            if (in.fail()) break;
+            if (auto pos = line.find("Hugepagesize:"); pos != std::string::npos) {
+                std::istringstream str(line.data() + pos + 13);
+                int aux;
+                str >> aux;
+                if (!str.fail()) pg_size = aux;
+                break;
+            }
+        }
+    }
+    return pg_size * 1024;
+}
+#endif
+
+llama_mmap::~llama_mmap() {
+    for (const auto & frag : mapped_fragments) {
+        if (munmap((char *) addr + frag.first, frag.second - frag.first)) {
+            LLAMA_LOG_WARN("warning: munmap failed: %s\n", strerror(errno));
+        }
+    }
+}
+
+#elif defined(_WIN32)
+
+llama_mmap::llama_mmap(struct llama_file * file, size_t prefetch, bool numa, [[maybe_unused]] bool use_thp) {
+    GGML_UNUSED(numa);
+
+    size = file->size;
+
+    HANDLE hFile = (HANDLE) _get_osfhandle(_fileno(file->fp));
+
+    HANDLE hMapping = CreateFileMappingA(hFile, NULL, PAGE_READONLY, 0, 0, NULL);
+
+    if (hMapping == NULL) {
+        DWORD error = GetLastError();
+        throw std::runtime_error(format("CreateFileMappingA failed: %s", llama_format_win_err(error).c_str()));
+    }
+
+    addr = MapViewOfFile(hMapping, FILE_MAP_READ, 0, 0, 0);
+    DWORD error = GetLastError();
+    CloseHandle(hMapping);
+
+    if (addr == NULL) {
+        throw std::runtime_error(format("MapViewOfFile failed: %s", llama_format_win_err(error).c_str()));
+    }
+
+    if (prefetch > 0) {
+#if _WIN32_WINNT >= 0x602
+        // PrefetchVirtualMemory is only present on Windows 8 and above, so we dynamically load it
+        BOOL (WINAPI *pPrefetchVirtualMemory) (HANDLE, ULONG_PTR, PWIN32_MEMORY_RANGE_ENTRY, ULONG);
+        HMODULE hKernel32 = GetModuleHandleW(L"kernel32.dll");
+
+        // may fail on pre-Windows 8 systems
+        pPrefetchVirtualMemory = reinterpret_cast<decltype(pPrefetchVirtualMemory)> (GetProcAddress(hKernel32, "PrefetchVirtualMemory"));
+
+        if (pPrefetchVirtualMemory) {
+            // advise the kernel to preload the mapped memory
+            WIN32_MEMORY_RANGE_ENTRY range;
+            range.VirtualAddress = addr;
+            range.NumberOfBytes = (SIZE_T) std::min(size, prefetch);
+            if (!pPrefetchVirtualMemory(GetCurrentProcess(), 1, &range, 0)) {
+                LLAMA_LOG_WARN("warning: PrefetchVirtualMemory failed: %s\n",
+                        llama_format_win_err(GetLastError()).c_str());
+            }
+        }
+#else
+        throw std::runtime_error("PrefetchVirtualMemory unavailable");
+#endif
+    }
+}
+
+void llama_mmap::unmap_fragment(size_t first, size_t last) {
+    // not supported
+    GGML_UNUSED(first);
+    GGML_UNUSED(last);
+}
+
+llama_mmap::~llama_mmap() {
+    if (!UnmapViewOfFile(addr)) {
+        LLAMA_LOG_WARN("warning: UnmapViewOfFile failed: %s\n",
+                llama_format_win_err(GetLastError()).c_str());
+    }
+}
+
+#else
+
+llama_mmap::llama_mmap(struct llama_file * file, size_t prefetch, bool numa, bool use_thp) {
+    GGML_UNUSED(file);
+    GGML_UNUSED(prefetch);
+    GGML_UNUSED(numa);
+    GGML_UNUSED(use_thp);
+
+    throw std::runtime_error("mmap not supported");
+}
+
+void llama_mmap::unmap_fragment(size_t first, size_t last) {
+    GGML_UNUSED(first);
+    GGML_UNUSED(last);
+
+    throw std::runtime_error("mmap not supported");
+}
+#endif
+using llama_mmaps = std::vector<std::unique_ptr<llama_mmap>>;
+
+
+void llama_mlock::init(void * ptr) {
+    GGML_ASSERT(addr == NULL && size == 0); // NOLINT
+    addr = ptr;
+}
+
+void llama_mlock::grow_to(size_t target_size) {
+    GGML_ASSERT(addr);
+    if (failed_already) {
+        return;
+    }
+    size_t granularity = lock_granularity();
+    target_size = (target_size + granularity - 1) & ~(granularity - 1);
+    if (target_size > size) {
+        if (raw_lock((uint8_t *) addr + size, target_size - size)) {
+            size = target_size;
+        } else {
+            failed_already = true;
+        }
+    }
+}
+
+#ifdef _POSIX_MEMLOCK_RANGE
+
+static size_t llama_m::lock_granularity() {
+    return (size_t) sysconf(_SC_PAGESIZE);
+}
+
+#ifdef __APPLE__
+    #define MLOCK_SUGGESTION \
+        "Try increasing the sysctl values 'vm.user_wire_limit' and 'vm.global_user_wire_limit' and/or " \
+        "decreasing 'vm.global_no_user_wire_amount'.  Also try increasing RLIMIT_MEMLOCK (ulimit -l).\n"
+#else
+    #define MLOCK_SUGGESTION \
+        "Try increasing RLIMIT_MEMLOCK ('ulimit -l' as root).\n"
+#endif
+
+bool llama_mlock::raw_lock(const void * addr, size_t size) const {
+    if (!mlock(addr, size)) {
+        return true;
+    }
+
+    char* errmsg = std::strerror(errno);
+    bool suggest = (errno == ENOMEM);
+
+    // Check if the resource limit is fine after all
+    struct rlimit lock_limit;
+    if (suggest && getrlimit(RLIMIT_MEMLOCK, &lock_limit)) {
+        suggest = false;
+    }
+    if (suggest && (lock_limit.rlim_max > lock_limit.rlim_cur + size)) {
+        suggest = false;
+    }
+
+    LLAMA_LOG_WARN("warning: failed to mlock %zu-byte buffer (after previously locking %zu bytes): %s\n%s",
+            size, this->size, errmsg, suggest ? MLOCK_SUGGESTION : "");
+    return false;
+}
+
+#undef MLOCK_SUGGESTION
+
+void llama_mlock::raw_unlock(void * addr, size_t size) {
+    if (munlock(addr, size)) {
+        LLAMA_LOG_WARN("warning: failed to munlock buffer: %s\n", std::strerror(errno));
+    }
+}
+
+#elif defined(_WIN32)
+
+size_t llama_mlock::lock_granularity() {
+    SYSTEM_INFO si;
+    GetSystemInfo(&si);
+    return (size_t) si.dwPageSize;
+}
+
+bool llama_mlock::raw_lock(void * ptr, size_t len) const {
+    for (int tries = 1; ; tries++) {
+        if (VirtualLock(ptr, len)) {
+            return true;
+        }
+        if (tries == 2) {
+            LLAMA_LOG_WARN("warning: failed to VirtualLock %zu-byte buffer (after previously locking %zu bytes): %s\n",
+                    len, size, llama_format_win_err(GetLastError()).c_str());
+            return false;
+        }
+
+        // It failed but this was only the first try; increase the working
+        // set size and try again.
+        SIZE_T min_ws_size, max_ws_size;
+        if (!GetProcessWorkingSetSize(GetCurrentProcess(), &min_ws_size, &max_ws_size)) {
+            LLAMA_LOG_WARN("warning: GetProcessWorkingSetSize failed: %s\n",
+                    llama_format_win_err(GetLastError()).c_str());
+            return false;
+        }
+        // Per MSDN: "The maximum number of pages that a process can lock
+        // is equal to the number of pages in its minimum working set minus
+        // a small overhead."
+        // Hopefully a megabyte is enough overhead:
+        size_t increment = len + 1048576;
+        // The minimum must be <= the maximum, so we need to increase both:
+        min_ws_size += increment;
+        max_ws_size += increment;
+        if (!SetProcessWorkingSetSize(GetCurrentProcess(), min_ws_size, max_ws_size)) {
+            LLAMA_LOG_WARN("warning: SetProcessWorkingSetSize failed: %s\n",
+                    llama_format_win_err(GetLastError()).c_str());
+            return false;
+        }
+    }
+}
+
+void llama_mlock::raw_unlock(void * ptr, size_t len) {
+    if (!VirtualUnlock(ptr, len)) {
+        LLAMA_LOG_WARN("warning: failed to VirtualUnlock buffer: %s\n",
+                llama_format_win_err(GetLastError()).c_str());
+    }
+}
+
+#else
+
+size_t llama_mlock::lock_granularity() {
+    return (size_t) 65536;
+}
+
+bool llama_mlock::raw_lock([[maybe_unused]] void * addr, [[maybe_unused]] size_t len) const {
+    LLAMA_LOG_WARN("warning: mlock not supported on this system\n");
+    return false;
+}
+
+void llama_mlock::raw_unlock([[maybe_unused]] void * addr, [[maybe_unused]] size_t len) {}
+
+#endif
